# film-manager-implementation/api/openapi.yaml
...
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: http://localhost:3001/
...
paths:
...
  "/api/films/private":
    get:
      tags:
      - api/films/private
      security:
      - cookieAuth: []
      summary: Retrieve the private films of the logged-in user
      description: The private films of the logged-in user are retrieved. A pagination
        mechanism is used to limit the size of messages.
      operationId: getPrivateFilms
      parameters:
      - in: query
        name: pageNo
        required: false
        schema:
          type: integer
        description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {}
...
  "/api/films/public":
    get:
      tags:
      - api/films/public
      summary: Retrieve the public films
      description: The public films (i.e., the films that are visible for all the
        users of the service) are retrieved. This operation does not require authentication.
        A pagination mechanism is used to limit the size of messages.
      operationId: getPublicFilms
      parameters:
      - in: query
        name: pageNo
        required: false
        schema:
          type: integer
        description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
        '500':
          description: Internal Server Error
          content: {}
  "/api/films/public/invited":
    get:
      tags:
      - api/films/public/invited
      security:
      - cookieAuth: []
      summary: Retrieve the public films that the logged-in user has been invited
        to review
      description: The public films that the logged-in user has been invited to review
        are retrieved. A pagination mechanism is used to limit the size of messages.
      operationId: getInvitedFilms
      parameters:
      - in: query
        name: pageNo
        required: false
        schema:
          type: integer
        description: The id of the requested page (if absent, the first page is returned)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {}
...
  "/api/users/authenticator":
    post:
      tags:
      - api/users/authenticator
      summary: Logs a user in
      description: The user who wants to log in sends the user data to the
        authenticator which performs the operation. If the request for the login of a new user
        comes from an already authenticated user, the previous user is first logged out.
      operationId: authenticateUser
      requestBody:
        description: The data of the user who wants to perform log in. The data
          structure must contain email and password.
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/User"
        required: true
      responses:
        '204':
          description: No Content
          headers:
            Set-Cookie:
              schema:
                type: string
                example: token=abcde12345; Path=/; HttpOnly
        '401':
          description: Unauthorized
          content: {}
...
components:
  schemas:
...
    User:
      description: schema for User data structures
      type: object
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            data structure describing the user
        id:
          type: integer
          description: unique identifier for the user
        email:
          type: string
          format: email
          description: user email
        name:
          type: string
          description: username
        password:
          type: string
          minLength: 6
          maxLength: 20
          description: user password
        self:
          type: string
          format: uri
      required:
      - email
...
    Film:
      description: schema for Film data structures
      type: object
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        id:
          type: integer
          description: unique identifier for the film data structure
        title:
          type: string
          description: textual title of the film
        owner:
          type: integer
          description: unique identifier of the the film data structure owner
        private:
          type: boolean
          description: true if only the owner of the film data structure can access
            it, false otherwise
          default: 'true'
        watchDate:
          type: string
          format: date
          description: the date when the film was watched by the owner
        rating:
          type: integer
          description: the rating the owner has given to the film
          minimum: 1
          maximum: 10
        favorite:
          type: boolean
          description: 'true if the film is among the favourite ones of the owner,
            false otherwise '
          default: 'false'
        self:
          type: string
          format: uri
        reviews:
          type: string
          format: uri
          description: the reviews for the film
      required:
      - title
      - owner
      - private
    Films:
      description: schema for Films data structures
      type: object
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            data structure describing the films
        totalPages:
          type: integer
          minimum: 0
          description: total number of pages
        currentPage:
          type: integer
          minimum: 0
          description: the page included in this object
        totalItems:
          type: integer
          minimum: 0
          description: the total number of items in the current page
        films:
          type: array
          items:
            "$ref": "#/components/schemas/Film"
          description: the films of the current page
        next:
          type: string
          format: uri
      required:
      - totalPages
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid